sys_clock: fix _sys_idle_elapsed_ticks type
Fixes the type of global variable _sys_idle_elapsed_ticks such that
both its nanokernel and microkernel definitions are of the same
type--int32_t.

One of the repercussions of this is that code related to the routine
_nano_sys_clock_tick_announce() does not need the application of
typecasting.

+++ kernel/microkernel/k_ticker.c
@@ -175,7 +175,7 @@ int _k_ticker(int event)
 		_TimeSliceUpdate();
 		_k_timer_list_update(ticks);
 		sys_clock_increment(ticks);
-		_nano_sys_clock_tick_announce((uint32_t)ticks);
+		_nano_sys_clock_tick_announce(ticks);
 	}
 
 	return 1;
